IJPS-20 refactored all packages and references from HMC namespace to JPS
