+++ .github/workflows/swagger.yaml
@@ -21,7 +21,7 @@ jobs:
         with:
           java-version: 17
       - name: Run Swagger Publisher
-        run: ./gradlew integration --tests uk.gov.hmcts.reform.hmc.config.SwaggerPublisherTest
+        run: ./gradlew integration --tests uk.gov.hmcts.reform.jps.config.SwaggerPublisherTest
       - name: Commit to repository
         run: |
           mkdir swagger-staging
