+++ src/integrationTest/java/uk/gov/hmcts/reform/jps/config/SwaggerPublisherTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.hmc.jp.config;
+package uk.gov.hmcts.reform.jps.config;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
