+++ src/main/resources/application.yaml
@@ -12,7 +12,7 @@ management:
         include: health, info, prometheus
 
 springdoc:
-  packagesToScan: uk.gov.hmcts.reform.hmc.jp.controllers
+  packagesToScan: uk.gov.hmcts.reform.jps.controllers
 
 #If you use a database then uncomment below lines and update db properties accordingly
 spring:
